usando jstl para carrergar tabela de usuarios e bloqueando usuario admin

+++ src/main/java/dao/DaoUsuariorepository.java
@@ -48,9 +48,30 @@ public class DaoUsuariorepository {
 
 	}
 	
+	
+	
+	public List<ModelLogin>consultarTodosUsuario() throws SQLException{
+		List<ModelLogin> lista = new ArrayList<>();
+		String sql = "select * from model_login where user_admin is false ";
+		PreparedStatement statement = connection.prepareStatement(sql);
+		ResultSet resultSet = statement.executeQuery();
+		while(resultSet.next()) {
+			ModelLogin usuario = new ModelLogin();
+			usuario.setEmail(resultSet.getString("email"));
+			usuario.setLogin(resultSet.getString("login"));
+			usuario.setId(resultSet.getLong("id"));
+			usuario.setNome(resultSet.getString("nome"));
+			//usuario.setSenha(resultSet.getString("senha"));			
+			lista.add(usuario);			
+		}
+		
+		
+		return lista;
+	}
+	
 	public List<ModelLogin>consultarUsuarioList(String nome) throws SQLException{
 		List<ModelLogin> lista = new ArrayList<>();
-		String sql = "select * from model_login where upper(nome) like upper(?)";
+		String sql = "select * from model_login where upper(nome) like upper(?) and  user_admin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setString(1, "%"+nome+"%");
 		ResultSet resultSet = statement.executeQuery();
@@ -70,7 +91,7 @@ public class DaoUsuariorepository {
 
 	public ModelLogin consultaUsuario(String login) throws SQLException {
 		ModelLogin modelLogin = new ModelLogin();
-		String sql = "select * from model_login where upper(login) =  upper ('" + login + "');";
+		String sql = "select * from model_login where upper(login) =  upper ('" + login + "') and  user_admin is false;";
 		PreparedStatement statement = connection.prepareStatement(sql);
 
 		ResultSet resultado = statement.executeQuery();
@@ -87,7 +108,7 @@ public class DaoUsuariorepository {
 	
 	public ModelLogin consultaUsuarioId(String id) throws SQLException {
 		ModelLogin modelLogin = new ModelLogin();
-		String sql = "select * from model_login where id = ?";
+		String sql = "select * from model_login where id = ? and  user_admin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setLong(1, Long.parseLong(id));
 
@@ -114,7 +135,7 @@ public class DaoUsuariorepository {
 	}
 
 	public void deletarUser(String idUser) throws SQLException {
-		String sql = "delete from model_login where id = ?";
+		String sql = "delete from model_login where id = ? and  user_admin is false";
 
 		PreparedStatement prepareSQL = connection.prepareStatement(sql);
